# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HALO — Archangel Operations Layer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: propertyId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateJob
      tags: [jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/complete:
    post:
      operationId: completeJob
      tags: [jobs]
      summary: Mark a job complete (sets warranty, opens ready-to-invoice)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/schedule:
    post:
      operationId: scheduleJob
      tags: [jobs]
      summary: Schedule a job for a date and crew leader
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "200":
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /crews:
    get:
      operationId: listCrews
      tags: [jobs]
      summary: Crew members with today's dispatch status
      responses:
        "200":
          description: Crews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrewToday"
    post:
      operationId: createCrew
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}/adjust:
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Adjust inventory quantity by a delta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjust"
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /vendors:
    get:
      operationId: listVendors
      tags: [inventory]
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [inventory]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [inventory]
      summary: Mark a purchase order received
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
...
  /notifications/{id}/read:
    post:
      operationId: readNotification
      tags: [activity]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /activities:
    get:
      operationId: listActivities
      tags: [activity]
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
        - name: entityId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Job:
      type: object
      required: [id, jobNo, status]
      properties:
        id: { type: string }
        jobNo: { type: string }
        woNo: { type: ["string", "null"] }
        propertyId: { type: string }
        propertyName: { type: ["string", "null"] }
        unitNo: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { type: string }
        crewLeaderId: { type: ["string", "null"] }
        crewLeaderName: { type: ["string", "null"] }
        bidId: { type: ["string", "null"] }
        contactId: { type: ["string", "null"] }
        inspectionRequired: { type: ["boolean", "null"] }
        inspectionPassedAt: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        recapSentAt: { type: ["string", "null"] }
        warrantyUntil: { type: ["string", "null"] }
        scheduledOn: { type: ["string", "null"] }
        grossProfit: { type: ["number", "null"] }
        marginPct: { type: ["number", "null"] }
        createdAt: { type: ["string", "null"] }
    JobDetail:
      type: object
      required: [job, activities, expenses, schedules]
      properties:
        job:
          $ref: "#/components/schemas/Job"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        schedules:
          type: array
          items:
            $ref: "#/components/schemas/Schedule"
    JobInput:
      type: object
      required: [propertyId, description]
      properties:
        propertyId: { type: string }
        woNo: { type: string }
        unitNo: { type: string }
        category: { type: string }
        description: { type: string, minLength: 1 }
        crewLeaderId: { type: string }
        contactId: { type: string }
        inspectionRequired: { type: boolean }
    JobUpdate:
      type: object
      properties:
        woNo: { type: string }
        unitNo: { type: string }
        category: { type: string }
        description: { type: string }
        status: { type: string }
        crewLeaderId: { type: string }
        inspectionRequired: { type: boolean }
    Schedule:
      type: object
      required: [id, jobId, scheduledOn, status]
      properties:
        id: { type: string }
        jobId: { type: string }
        scheduledOn: { type: string }
        windowStart: { type: ["string", "null"] }
        crewLeaderId: { type: ["string", "null"] }
        status: { type: string }
    ScheduleInput:
      type: object
      required: [scheduledOn]
      properties:
        scheduledOn: { type: string }
        windowStart: { type: string }
        crewLeaderId: { type: string }
    Crew:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        trade: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        isLeader: { type: ["boolean", "null"] }
        active: { type: ["boolean", "null"] }
    CrewToday:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        trade: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        isLeader: { type: ["boolean", "null"] }
        todayStatus: { type: ["string", "null"], description: "route | site | done | idle" }
        todayJob: { type: ["string", "null"] }
        todayProperty: { type: ["string", "null"] }
    CrewInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        trade: { type: string }
        phone: { type: string }
        isLeader: { type: boolean }
...
    Expense:
      type: object
      required: [id, amount]
      properties:
        id: { type: string }
        jobId: { type: ["string", "null"] }
        propertyId: { type: ["string", "null"] }
        vendor: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        amount: { type: number }
        source: { type: ["string", "null"] }
        spentOn: { type: ["string", "null"] }
...
    InventoryItem:
      type: object
      required: [id, name, qty, reorderAt]
      properties:
        id: { type: string }
        name: { type: string }
        qty: { type: number }
        reorderAt: { type: number }
        unitCost: { type: ["number", "null"] }
        preferredVendor: { type: ["string", "null"] }
        low: { type: ["boolean", "null"] }
    InventoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        qty: { type: number }
        reorderAt: { type: number }
        unitCost: { type: number }
        preferredVendor: { type: string }
    InventoryAdjust:
      type: object
      required: [delta]
      properties:
        delta: { type: number }
        reason: { type: string }
    Vendor:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        trade: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        coiExpiresOn: { type: ["string", "null"] }
        compliant: { type: ["boolean", "null"] }
    VendorInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        trade: { type: string }
        email: { type: string }
        phone: { type: string }
        coiExpiresOn: { type: string }
    PurchaseOrder:
      type: object
      required: [id, poNo, status]
      properties:
        id: { type: string }
        poNo: { type: string }
        vendorId: { type: ["string", "null"] }
        vendorName: { type: ["string", "null"] }
        jobId: { type: ["string", "null"] }
        jobNo: { type: ["string", "null"] }
        expectedOn: { type: ["string", "null"] }
        receivedAt: { type: ["string", "null"] }
        status: { type: string }
        late: { type: ["boolean", "null"] }
    PurchaseOrderInput:
      type: object
      required: [vendorId]
      properties:
        vendorId: { type: string }
        jobId: { type: string }
        expectedOn: { type: string }
    Notification:
      type: object
      required: [id, kind, priority, title]
      properties:
        id: { type: string }
        kind: { type: string }
        priority: { type: string }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["string", "null"] }
        title: { type: string }
        body: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"] }
        readAt: { type: ["string", "null"] }
        resolvedAt: { type: ["string", "null"] }
    Activity:
      type: object
      required: [id, entityType, entityId, kind]
      properties:
        id: { type: string }
        entityType: { type: string }
        entityId: { type: string }
        kind: { type: string }
        body: { type: ["string", "null"] }
        storagePath: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
    ActivityInput:
      type: object
      required: [entityType, entityId, kind]
      properties:
        entityType: { type: string }
        entityId: { type: string }
        kind: { type: string }
        body: { type: string }
...