# lib/api-zod/src/generated/types/crew.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * HALO — Archangel Operations Layer API
 * OpenAPI spec version: 0.1.0
 */

export interface Crew {
  id: string;
  name: string;
  /** @nullable */
  trade?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  isLeader?: boolean | null;
  /** @nullable */
  active?: boolean | null;
}