# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * HALO — Archangel Operations Layer API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: string;
  /** @nullable */
  jobId?: string | null;
  /** @nullable */
  propertyId?: string | null;
  /** @nullable */
  vendor?: string | null;
  /** @nullable */
  category?: string | null;
  amount: number;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  spentOn?: string | null;
}