# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * HALO — Archangel Operations Layer API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Expense } from './expense';
import type { Job } from './job';
import type { Schedule } from './schedule';

export interface JobDetail {
  job: Job;
  activities: Activity[];
  expenses: Expense[];
  schedules: Schedule[];
}