# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * HALO — Archangel Operations Layer API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleInput {
  scheduledOn: string;
  windowStart?: string;
  crewLeaderId?: string;
}